pyproject.toml
[project]
name = "s2dm"
dynamic = ["version"]
description = "An friendly approach for modeling semantic data ."
authors = [
    { name = "Daniel Alvarez-Coello", email = "8550265+jdacoello@users.noreply.github.com" },
]
license-files = ["LICENSE"]
readme = "README.md"
dependencies = [
    "click>=8.1.7",
    "rich-click>=1.8.3",
    "rich>=13.9.4",
    "graphql-core>=3.2.6",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.6",
    "jinja2>=3.1.0",
    "rdflib>=7.1.3",
    "pyshacl>=0.30.0",
    "ariadne>=0.24.0",
    "langcodes>=3.5.0",
    "case-converter>=1.2.0",
    "requests>=2.32.0",
]
requires-python = ">=3.11"

[project.urls]
homepage = "https://github.com/COVESA/s2dm"
documentation = "https://github.com/COVESA/s2dm/blob/main/README.md"

[project.scripts]
s2dm = "s2dm.cli:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/s2dm/__init__.py"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.13.0",
    "ruff>=0.7.3",
    "pre-commit>=4.0.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
    "nox>=2024.10.9",
    "bump-my-version>=0.28.1",
    "faker>=37.1.0",
    "hypothesis>=6.131.9",
    "types-pyyaml>=6.0.12.20250516",
    "gitlint>=0.20.0.dev48",
]

[tool.mypy]
files = ["src/s2dm"]
show_error_codes = true
strict = true
enable_error_code = ["redundant-expr", "truthy-bool"]
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov --cov-report=xml:coverage.xml"


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PGH", "B", "SIM"]
